refactor: tasks↔taskRunner の循環依存を解消
コマンドパレット登録を leaf モジュール (commands/taskCommands.ts) に
切り出し、tasks ストアから taskRunner / commandStore への依存を除去。
未使用だった TASK_COMMAND_PREFIX の再エクスポートも削除。
